deploy: Remove interactive prompts and some side effects

+++ deploy.sh
@@ -7,13 +7,13 @@ files=(emacs.d gitconfig gtkrc-2.0 msmtprc muttrc offlineimaprc \
        quiltrc tmux.conf vimrc xbindkeysrc xinitrc xmonad Xresources zshrc)
 
 for file in $files; do;
-  ln -isT "$pwd/$file" "$HOME/.$file"
+  ln -fsT "$pwd/$file" "$HOME/.$file"
 done
 
 mkdir -p "$HOME/.config/gtk-3.0"
-ln -is "$pwd/settings.ini" "$HOME/.config/gtk-3.0/settings.ini"
+ln -fs "$pwd/settings.ini" "$HOME/.config/gtk-3.0/settings.ini"
 
-ln -isT "$HOME/.xinitrc" "$HOME/.xsession"
+ln -fsT "$HOME/.xinitrc" "$HOME/.xsession"
 
 mkdir -p "$HOME/msg/.offlineimap"
 echo '#!/bin/sh\nofflineimap' >! "$HOME/msg/.offlineimap/run"
@@ -23,27 +23,33 @@ chmod +x "$HOME/msg/.offlineimap/run"
 mkdir -p "$HOME/bin"
 find "$pwd/bin" -type f -exec basename '{}' \; |
   while read cmd; do;
-    ln -isT "$pwd/bin/$cmd" "$HOME/bin/$cmd"
+    ln -fsT "$pwd/bin/$cmd" "$HOME/bin/$cmd"
   done
 
 # For now, we need to build dzen from source, since the version in the
 # Debian repositories is rather outdated.
-svn checkout http://dzen.googlecode.com/svn/trunk/ dzen-bin
-cd dzen-bin
-sudo make clean install
-cd gadgets
-sudo make clean install
-cd $pwd
-rm -rf dzen-bin
+if ( ! command -v dzen2 > /dev/null ); then
+  svn checkout http://dzen.googlecode.com/svn/trunk/ dzen-bin
+  cd dzen-bin
+  sudo make clean install
+  cd gadgets
+  sudo make clean install
+  cd $pwd
+  rm -rf dzen-bin
+fi
 
 # Let's get some icons for use in our dzen display
-wget --quiet http://bitbucket.org/jerronymous/dotfiles/src/d8b5855a5ec6/.xmonad/dzen/icon-packs/xbm8x8-0.1.tar.gz
-tar -xf xbm8x8-0.1.tar.gz
-mkdir -p dzen/icons
-mv xbm8x8/* dzen/icons
-rm -rf xbm8x8*
+if [ ! -d "$pwd/dzen/icons" ]; then;
+  wget --quiet http://bitbucket.org/jerronymous/dotfiles/src/d8b5855a5ec6/.xmonad/dzen/icon-packs/xbm8x8-0.1.tar.gz
+  tar -xf xbm8x8-0.1.tar.gz
+  mkdir -p dzen/icons
+  mv xbm8x8/* dzen/icons
+  rm -rf xbm8x8*
+fi
 
 # Before we can install packages for Emacs23, we need a package manager
-wget --quiet -O "$pwd/emacs.d/package.el" \
-  $(wget --quiet -O - http://marmalade-repo.org/ | \
-      sed -n 's|.*\(http://.*package.el\)">|\1|p')
+if [ ! -f "$pwd/emacs.d/package.el" ]; then
+  wget --quiet -O "$pwd/emacs.d/package.el" \
+    $(wget --quiet -O - http://marmalade-repo.org/ | \
+        sed -n 's|.*\(http://.*package.el\)">|\1|p')
+fi
